feat :: logout api connenct
