+++ src/all/component/menu/items.ts
@@ -1,7 +1,7 @@
 export const items = [
     { title: "물품 신청 하러가기", path: "/object" },
     { title: "공지사항 보러가기", path: "/notice" },
-    // { title: "월말평가 작성하기", path: "/" },
+    { title: "월말평가 작성하기", path: "/" },
     { title: "팀스페이스 보러가기", path: "/team-space" },
     { title: "역대 전공동아리 보러가기", path: "/club-history" },
 ];
