+++ src/all/component/modal/Setting/style.ts
