+++ src/api/login/login.ts
@@ -25,3 +25,13 @@ export default async function googleLogin(code: string): Promise<LoginResponse |
     return null;
   }
 }
+
+export async function logout(refreshToken: string) {
+  try {
+    const res = await axiosInstance.post('/ara/auth/logout', { refreshToken });
+    return res.status === 200;
+  } catch (err) {
+    console.error('로그아웃 실패', err);
+    return false;
+  }
+}
